save company

+++ laravel/app/Services/User/ProfileService.php
@@ -30,62 +30,42 @@ class ProfileService
         return $validator;
     }
 
-//
+
     public function createCompany(Request $request)
     {
-//        $this->db->insert(
-//            'agt_comp_items',
-//            ['author_id' => $this->id,
-//                'topic_id' => 0, 'type_id' => 0,
-//                'obl_id' => $region, 'ray_id' => 0,
-//                'title' => $title, 'title_full' => '',
-//                'city' => $city, 'phone' => '',
-//                'phone2' => '', 'phone3' => '',
-//                'email' => '', 'www' => '',
-//                'zipcode' => $zipcode,
-//                'addr' => $addr,
-//                'add_date' => 'now()',
-//                'content' => $content,
-//                'short' => $compshort,
-//                'contacts' => '',
-//                'logo_file' => $filename]);
-//        $compId = $this->db->getLastId();
-        //$data = $request->all();
-        //$data = $this->createCompanyValidator($data);
-
-
-        if(!empty($request->all())){
-        /** TODO id поменять на user_id */
-        $author_id = auth()->user()->user_id;
-        $file = $request->file('logo');
-        $fileName = '';
-
-        if($file->getError() == 0)
+        /** var User $user */
+        if($request->isMethod('post'))
         {
-            $fileName = $file->getClientOriginalName();
-            $file->move('/var/www/agrotender'.self::PART_FILE_NAME, $fileName);
-        }
+            $author_id = auth()->user()->user_id;
+            $file = $request->file('logo');
+            $fileName = '';
 
-        $compshort = strlen($request->get('content')) > 210 ? Str::limit($request->get('content'), 200) : $request->get('content');
-
-        $company = CompItems::updateOrCreate(['author_id' => $author_id], $request->except(['_token', 'logo']) + [
-            'author_id' => $author_id, 'topic_id' => 0, 'type_id' => 0,
-            'ray_id' => 0, 'title_full' => '', 'phone' => '', 'short' => $compshort,
-            'phone2' => '', 'phone3' => '', 'www' => '', 'add_date' => Carbon::now()->toDateTimeString(),
-            'contacts' => '', 'logo_file' => self::PART_FILE_NAME.$fileName
-        ], $request->toArray());
-
-        CompTopicItem::where('item_id', $company->id)->delete();
-
-        \DB::beginTransaction();
-            foreach ($request->get('rubrics') as $index => $rubric) {
-                CompTopicItem::create([
-                    'topic_id' => (int)$rubric,
-                    'item_id' => $company->id,
-                    'add_date' => Carbon::now()->toDateTimeString()
-                ]);
+            if($file && $file->getError() == 0)
+            {
+                $fileName = $file->getClientOriginalName();
+                $file->move('/var/www/agrotender'.self::PART_FILE_NAME, $fileName);
             }
-        \DB::commit();
+
+            $compshort = strlen($request->get('content')) > 210 ? Str::limit($request->get('content'), 200) : $request->get('content');
+
+            $company = CompItems::updateOrCreate(['author_id' => $author_id], $request->except(['_token', 'logo']) + [
+                'author_id' => $author_id, 'topic_id' => 0, 'type_id' => 0,
+                'ray_id' => 0, 'title_full' => '', 'phone' => '', 'short' => $compshort,
+                'phone2' => '', 'phone3' => '', 'www' => '', 'add_date' => Carbon::now()->toDateTimeString(),
+                'contacts' => '', 'logo_file' => self::PART_FILE_NAME.$fileName
+            ], $request->toArray());
+
+            CompTopicItem::where('item_id', $company->id)->delete();
+
+            \DB::beginTransaction();
+                foreach ($request->get('rubrics') as $index => $rubric) {
+                    CompTopicItem::create([
+                        'topic_id' => (int)$rubric,
+                        'item_id' => $company->id,
+                        'add_date' => Carbon::now()->toDateTimeString()
+                    ]);
+                }
+            \DB::commit();
         }
     }
 }
